enhance: tighten FFI validation and writer schema errors
This tightens FFI-side argument validation so bad inputs are caught consistently instead of slipping into asserts, null dereferences, or silent defaults.

Writer FFI paths now import record batches with the writer's declared schema, so schema mismatches between the writer and incoming record batch are returned as structured FFI errors instead of surfacing later in the write path.

This also makes delta log num_entries use int64_t across the C and Python FFI surface, rejects non-positive counts, and returns proper status errors from the bridge layer when imported manifest data is malformed.

+++ cpp/include/milvus-storage/ffi_c.h
@@ -207,7 +207,7 @@ typedef struct LoonColumnGroups {
  */
 typedef struct LoonDeltaLogs {
   const char** delta_log_paths;
-  uint32_t* delta_log_num_entries;
+  int64_t* delta_log_num_entries;
   uint32_t num_delta_logs;
 } LoonDeltaLogs;
